requirements.txt
aiofiles==22.1.0
anyio==3.6.1
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
datetime-utils==1.0.0
decouple==0.0.7
dnspython==2.2.1
enum34==1.1.10
fastapi==0.92.0
fastapi-health==0.4.0
h11==0.13.0
idna==3.3
loguru==0.6.0
motor==3.0.0
pdf2image==1.16.0
Pillow==9.3.0
py-healthcheck==1.10.1
pydantic==1.9.2
PyJWT==2.6.0
pymongo==4.2.0
PyPDF2==2.10.9
python-dateutil==2.8.2
python-decouple==3.4
python-multipart==0.0.5
pytz==2022.6
qrcode==7.3.1
requests==2.28.1
schedule==1.2.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.2
win32-setctime==1.1.0
